+++ libs/RadioHead/RH_E32.cpp
@@ -57,7 +57,7 @@ bool RH_E32::reset()
 {
   setOperatingMode(ModeSleep);
   uint8_t resetCommand[] = { RH_E32_COMMAND_RESET, RH_E32_COMMAND_RESET, RH_E32_COMMAND_RESET };
-  size_t result = _s->write(resetCommand, sizeof(resetCommand));
+  int32_t result = _s->write(resetCommand, sizeof(resetCommand));
   setOperatingMode(ModeNormal);
   return (result == sizeof(resetCommand));
 }
@@ -67,7 +67,7 @@ bool RH_E32::readParameters(Parameters& params)
   setOperatingMode(ModeSleep);
   uint8_t readParamsCommand[] = { RH_E32_COMMAND_READ_PARAMS, RH_E32_COMMAND_READ_PARAMS, RH_E32_COMMAND_READ_PARAMS };
   _s->write(readParamsCommand, sizeof(readParamsCommand));
-  size_t result = _s->readBytes((char*)&params, sizeof(params)); // default 1 sec timeout
+  int32_t result = _s->readBytes((char*)&params, sizeof(params)); // default 1 sec timeout
   setOperatingMode(ModeNormal);
   return (result == sizeof(Parameters));
 }
@@ -77,7 +77,7 @@ bool RH_E32::writeParameters(Parameters& params, bool save)
   setOperatingMode(ModeSleep);
   params.head = save ? RH_E32_COMMAND_WRITE_PARAMS_SAVE : RH_E32_COMMAND_WRITE_PARAMS_NOSAVE;
   //  printBuffer("writing now", (uint8_t*)&params, sizeof(params));
-  size_t result = _s->write((uint8_t*)&params, sizeof(params));
+  int32_t result = _s->write((uint8_t*)&params, sizeof(params));
   if (result != sizeof(params))
     return false;
   
@@ -129,7 +129,7 @@ bool RH_E32::getVersion()
   uint8_t readVersionCommand[] = { RH_E32_COMMAND_READ_VERSION, RH_E32_COMMAND_READ_VERSION, RH_E32_COMMAND_READ_VERSION };
   _s->write(readVersionCommand, sizeof(readVersionCommand));
   uint8_t version[4];
-  size_t result = _s->readBytes((char *)version, sizeof(version)); // default 1 sec timeout
+  int32_t result = _s->readBytes((char *)version, sizeof(version)); // default 1 sec timeout
   setOperatingMode(ModeNormal);
   if (result == 4)
     {
