+++ src/lmic/hal.h
@@ -74,14 +74,14 @@ extern "C"
      *   - write the command byte 'cmd'
      *   - write 'len' bytes out of 'buf'
      */
-    void hal_spi_write(u1_t cmd, const u1_t *buf, size_t len);
+    void hal_spi_write(u1_t cmd, const u1_t *buf, int32_t len);
 
     /*
      * Perform SPI read transaction with radio chip
      *   - write the command byte 'cmd'
      *   - read 'len' bytes into 'buf'
      */
-    void hal_spi_read(u1_t cmd, u1_t *buf, size_t len);
+    void hal_spi_read(u1_t cmd, u1_t *buf, int32_t len);
 
     /*
      * disable all CPU interrupts.
