+++ src/lmic/oslmic.h
@@ -290,7 +290,7 @@ u2_t os_crc16(xref2cu1_t d, uint len);
 // progmem using pgm_read_xx, or accesses memory directly when the
 // index is a constant so gcc can optimize it away;
 #define TABLE_GETTER(postfix, type, pgm_type)                                                                          \
-    static inline type table_get##postfix(const type *table, size_t index)                                             \
+    static inline type table_get##postfix(const type *table, int32_t index)                                             \
     {                                                                                                                  \
         if (__builtin_constant_p(table[index]))                                                                        \
             return table[index];                                                                                       \
@@ -311,31 +311,31 @@ TABLE_GETTER(_ostime, ostime_t, dword);
 // For AVR, store constants in PROGMEM, saving on RAM usage
 #define CONST_TABLE(type, name) const type PROGMEM RESOLVE_TABLE(name)
 #else
-static inline u1_t table_get_u1(const u1_t *table, size_t index)
+static inline u1_t table_get_u1(const u1_t *table, int32_t index)
 {
     return table[index];
 }
-static inline s1_t table_get_s1(const s1_t *table, size_t index)
+static inline s1_t table_get_s1(const s1_t *table, int32_t index)
 {
     return table[index];
 }
-static inline u2_t table_get_u2(const u2_t *table, size_t index)
+static inline u2_t table_get_u2(const u2_t *table, int32_t index)
 {
     return table[index];
 }
-static inline s2_t table_get_s2(const s2_t *table, size_t index)
+static inline s2_t table_get_s2(const s2_t *table, int32_t index)
 {
     return table[index];
 }
-static inline u4_t table_get_u4(const u4_t *table, size_t index)
+static inline u4_t table_get_u4(const u4_t *table, int32_t index)
 {
     return table[index];
 }
-static inline s4_t table_get_s4(const s4_t *table, size_t index)
+static inline s4_t table_get_s4(const s4_t *table, int32_t index)
 {
     return table[index];
 }
-static inline ostime_t table_get_ostime(const ostime_t *table, size_t index)
+static inline ostime_t table_get_ostime(const ostime_t *table, int32_t index)
 {
     return table[index];
 }
